Fixed validateDay function
The bot doesn't forbid ordering goods to next year, but previously it
didn't notify the user, that the goods were ordered to the next year.
Now if user picks a months, that is less than current month
(according to numbers in Gregorian calendar) bot puts the order to
next year and notifies user about it.